Fix remaining decimal type conversions across the codebase

+++ Poliedro.Eds.Application/DashboardPowerBI/Dtos/ShoppingDto.cs
@@ -7,5 +7,5 @@ public record ShoppingDto
     DateTime Date,
     string IdProvider,
     string IdCategory,
-    double Amount
+    decimal Amount
     );
