+++ Poliedro.Eds.Infraestructure.Persistence.Mysql/Court/Repositories/CourtTransactionalService.cs
@@ -300,7 +300,7 @@ namespace Poliedro.Eds.Infraestructure.Persistence.Mysql.Court.Repositories
         {
             try
             {
-                var priceUpdates = new List<(int ProductId, double OldPrice, double NewPrice, string ProductName)>();
+                var priceUpdates = new List<(int ProductId, decimal OldPrice, double NewPrice, string ProductName)>();
 
                 foreach (var dispenser in courtDispensers)
                 {
@@ -332,11 +332,11 @@ namespace Poliedro.Eds.Infraestructure.Persistence.Mysql.Court.Repositories
                     var currentSellPrice = product.SellPrice ?? 0;
 
                     // Validar si el precio es diferente (usando tolerancia para comparación de decimales)
-                    if (Math.Abs(currentSellPrice - courtPrice) > 0.01)
+                    if (Math.Abs(currentSellPrice - (decimal)courtPrice) > 0.01m)
                     {
                         // Actualizar el precio del producto usando domain service
                         var previousPrice = product.SellPrice;
-                        product.SellPrice = courtPrice;
+                        product.SellPrice = (decimal)courtPrice;
                         
                         var updateResult = await productUpdateService.UpdateAsync(product);
                         if (!updateResult.IsSuccess)
@@ -392,7 +392,7 @@ namespace Poliedro.Eds.Infraestructure.Persistence.Mysql.Court.Repositories
                 var product = productResult.Value!;
 
                 // Validar que hay suficiente stock
-                var nuevoStock = product.Stock - dispenser.GallonsDifferenceResult;
+                var nuevoStock = product.Stock - (decimal)dispenser.GallonsDifferenceResult;
                 //if (nuevoStock < 0)
                 //{
                 //    return Error.BadRequest("InsufficientStock",
